Start roulette package by mocking math/rand
In fact, I'm not mocking it, just understanding the values it gives for
the first 10 so that, with proper knowledge, I can determine the exact
outputs of certain algorithms.

In essence, the roulette package (which will build biased
coins/dice/roulette wheels from fair ones) will take a rand.Random
parameter from which to draw randomness. I can now use my own and
guarantee the values produced, for testing.