Fix indexing error introduced by last commit
Now that we get proper sides of the dice (i.e. >0) I need to adjust the
indexing to map 1 -> 0, 2 -> 1, &c. A simple subtraction suffices.

I also noticed that the LoadedDie suffers from the same off-by-one in
the side that it returns, so an addition fixes that.

Finally, repeated runs of voseInit were changing the internals of the
die because I had failed to properly create a slice pointing to
different memory. Corrected.

I should consider caching the voseInit stage inside the die itself, as
it can be private and should be just a bit of refactoring to use the
pre-processing as intended.